refactoring + examples to the readthedocs
