+++ batcore/baselines/models/CN.py
@@ -6,7 +6,7 @@ from scipy.sparse import dok_matrix
 import networkx as nx
 from community import community_louvain
 
-from batcore.RecommenderBase.recommender import BanRecommenderBase
+from batcore.modelbase.recommender import BanRecommenderBase
 
 
 # TODO look into multiple owners
