+++ batcore/baselines/models/RevFinder.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 
 import numpy as np
 
-from batcore.RecommenderBase.recommender import BanRecommenderBase
+from batcore.modelbase.recommender import BanRecommenderBase
 from ..utils import LCSubseq, LCSubstr, LCSuff, LCP
 from ..utils import get_map
