+++ batcore/baselines/models/xFinder.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 import numpy as np
 
-from batcore.RecommenderBase.recommender import BanRecommenderBase
+from batcore.modelbase.recommender import BanRecommenderBase
 
 
 class xFinder(BanRecommenderBase):
