+++ batcore/counter/CounterBase.py
