+++ batcore/modelbase/__init__.py
@@ -1,0 +1,7 @@
+from .recommender import RecommenderBase
+from .recommender import BanRecommenderBase
+
+__all__ = [
+    "RecommenderBase",
+    "BanRecommenderBase",
+]
